projects/prompt-testing-framework/requirements.txt
# P4 — Prompt Testing Framework
# pip install -r requirements.txt
# or: pip install -e .

openai>=1.30.0
anthropic>=0.28.0

pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0

# Optional: for YAML test files
pyyaml>=6.0

jupyter>=1.0.0
ipython>=8.0.0
